Drop the stale legacy deferred-header test and clear compile warnings.
The v6 snapshot layout no longer matches the four-byte strip fixture.
Unused locals/helpers and host %3lu vs uint32_t were warning-only.

+++ src/Utils/DisplayWindowUtils.cpp
@@ -270,7 +270,7 @@ void formatLoopLengthBars(char* out, size_t outSize, uint32_t loopLengthTicks,
   if (bars > 999UL) {
     bars = 999UL;
   }
-  snprintf(out, outSize, "%3lu", bars);
+  snprintf(out, outSize, "%3lu", static_cast<unsigned long>(bars));
 }
 
 }  // namespace DisplayWindowUtils
